[JS] 탈퇴

+++ HoliDiary/src/main/webapp/resources/kjs_js/kjs_mypage.js
@@ -1,3 +1,7 @@
 function deleteUser(){
-	location.href="delete.user";
+	let ok = confirm('진짜로 삭제?');
+	
+	if(ok){
+		location.href="delete.user";
+	}
 }
